Implimenting Groups Functionality
I was talking to my mom on my way back to school and she mentioned what a pain it was to track how much money everyone owes during a trip. The idea behind groups is that you will make a group with 'n' people, then you can make new transactions and chose who is going to help pay for that particualar transation. Then there will be a calculate button or it will auto do it where is shows who owes who money. I implimented what I think will work but we should talk about this in class. Just stole code from the details screen

//Todo
I broke the ExpandedProfileDebug vertical scroll functionality. im calling it in our ExpandedGroupDebug and it doesn't like having a scrolling list in a scrolling list.

The code didnt complile, something wrong with the db. I commented out the repository so not it compiles, idk but hopefully its a quick fix